Create TaskList class and move task related activities.
